requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.5
django-allauth==0.55.2
django-annoying==0.10.6
django-appconf==1.0.5
django-avatar==7.1.1
django-comments-xtd==2.9.10
django-contrib-comments==2.2.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-dynamic-preferences==1.15.0
django-extensions==3.2.3
django-facebook==6.0.3
django-model-utils==4.3.1
django-querysetsequence==0.17
django-rest-multiple-models==2.1.3
django-static-precompiler==2.4
django-tgbot==2.0.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-serializer-extensions==2.0.1
dnspython==2.4.2
docutils==0.20.1
fcm-django==2.0.0
firebase-admin==5.4.0
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
hashids==1.3.1
httplib2==0.22.0
idna==3.4
jsonfield==3.1.0
msgpack==1.0.6
oauthlib==3.2.2
persisting-theory==1.0
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.24.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.16
